Attempt to fix build break.

+++ source/MaterialXGenShader/Syntax.cpp
@@ -113,7 +113,7 @@ string Syntax::getSwizzledVariable(const string& srcName, const TypeDesc* srcTyp
         else
         {
             int channelIndex;
-            if (channelIndex >= srcMembers.size() || (channelIndex = srcType->getChannelIndex(ch)) < 0)
+            if (channelIndex >= static_cast<int>(srcMembers.size()) || (channelIndex = srcType->getChannelIndex(ch)) < 0)
             {
                 throw ExceptionShaderGenError("Given channel index: '" + string(1,ch) + "' in channels pattern is incorrect for type '" + srcType->getName() + "'.");
             }
